db: relocate address-store residuals and split unit tests
Final cleanup of the address-store db-ops migration. The monolithic
addresses_common.go and addresses_common_test.go are removed.
